requirements.txt
absl-py==2.1.0
accelerate==1.1.1
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.14.0
altgraph==0.17.4
annotated-types==0.6.0
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.2.0
azure-core==1.30.1
azure-cosmos==4.7.0
azure-identity==1.16.0
Babel==2.15.0
banal==1.0.6
base==0.0.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
catboost==1.2.7
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.1.0
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
cryptography==42.0.7
cycler==0.12.1
dataclasses-json==0.6.6
dataset==1.6.2
datasets==3.1.0
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.8
distro==1.9.0
dnspython==2.6.1
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.16.0
Flask==3.0.3
Flask-Cors==5.0.0
flatbuffers==24.3.25
fonttools==4.53.1
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.9.0
gast==0.6.0
git-filter-repo==2.45.0
google-pasta==0.2.0
graphviz==0.20.3
greenlet==3.0.3
grpcio==1.66.1
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.6
idna==3.7
imbalanced-learn==0.12.3
imblearn==0.0
iniconfig==2.0.0
ipykernel==6.29.4
ipython==8.24.0
ipywidgets==8.1.2
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
json5==0.9.25
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter_server==2.14.0
jupyter_server_terminals==0.5.3
jupyterlab==4.2.0
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.1
jupyterlab_widgets==3.0.10
keras==3.5.0
kiwisolver==1.4.7
langchain==0.2.0
langchain-core==0.2.1
langchain-text-splitters==0.2.0
langsmith==0.1.60
libclang==18.1.1
lightgbm==4.5.0
llvmlite==0.43.0
Mako==1.3.6
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.4.1
mpmath==1.3.0
msal==1.28.0
msal-extensions==1.1.0
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
namex==0.0.8
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
nltk==3.9.1
notebook==7.2.0
notebook_shim==0.2.4
numba==0.60.0
numpy==1.26.4
openai==1.30.1
opencv-python==4.10.0.84
opt-einsum==3.3.0
optree==0.12.1
orjson==3.10.3
overrides==7.7.0
packaging==23.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pefile==2023.2.7
pillow==10.4.0
platformdirs==4.2.2
plotly==5.24.1
pluggy==1.5.0
portalocker==2.8.2
prometheus_client==0.20.0
prompt-toolkit==3.0.43
protobuf==4.25.5
psutil==5.9.8
pure-eval==0.2.2
pyarrow==18.1.0
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.18.0
pyinstaller==6.7.0
pyinstaller-hooks-contrib==2024.6
PyJWT==2.8.0
pymongo==4.7.2
pyparsing==3.1.4
pytest==8.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
pytz==2024.2
pywin32==306
pywin32-ctypes==0.2.2
pywinpty==2.0.13
PyYAML==6.0.1
pyzmq==26.0.3
qtconsole==5.5.2
QtPy==2.4.1
referencing==0.35.1
regex==2024.9.11
requests==2.32.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.8.1
rpds-py==0.18.1
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
Send2Trash==1.8.3
setuptools==70.0.0
shap==0.46.0
simplejson==3.19.2
six==1.16.0
slicer==0.0.8
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==1.4.54
stack-data==0.6.3
sympy==1.13.2
tenacity==8.3.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow_intel==2.18.0
termcolor==2.4.0
terminado==0.18.1
tf_keras==2.18.0
threadpoolctl==3.5.0
tinycss2==1.3.0
tokenizers==0.19.1
torch==2.4.1
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
transformers==4.44.2
types-python-dateutil==2.9.0.20240316
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.2
uri-template==1.3.0
urllib3==2.2.1
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.4
wheel==0.44.0
widgetsnbextension==4.0.10
wrapt==1.16.0
xgboost==2.1.2
xxhash==3.5.0
yarl==1.9.4
